+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,33 @@
 # smartagentkit
 
+## 0.1.10
+
+### Patch Changes
+
+- Updated dependencies [e701708]
+  - @smartagentkit/sdk@0.1.10
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [75078c2]
+  - @smartagentkit/sdk@0.1.9
+
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartagentkit/sdk@0.1.8
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies
+  - @smartagentkit/sdk@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes